api and datetime methods

+++ MyMap/src/main/java/space/harbour/mymap/domain/Hourly.java
@@ -3,9 +3,11 @@ package space.harbour.mymap.domain;
 import lombok.Data;
 import lombok.ToString;
 
+import java.util.List;
+
 @Data
 @ToString
 public class Hourly {
 
-    private List<double> time;
+    private List<Long> time;
 }
